app/requirements.txt
anyio==3.6.2
argcomplete==3.0.5
astroid==2.15.4
attrs==23.1.0
black==23.3.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
datamodel-code-generator==0.18.0
dill==0.3.6
dnspython==2.3.0
email-validator==2.0.0.post1
fastapi==0.95.1
genson==1.2.2
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
importlib-resources==5.12.0
inflect==5.6.2
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
jsonschema==4.17.3
jsonschema-spec==0.1.4
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
openapi-schema-validator==0.3.4
openapi-spec-validator==0.5.1
packaging==23.1
pathable==0.4.3
pathspec==0.11.1
platformdirs==3.2.0
pluggy==1.0.0
prance==0.22.2.22.0
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg-pool==3.1.7
pydantic==1.10.7
pylint==2.17.3
pyrsistent==0.19.3
PySnooper==1.1.1
pytest==7.3.1
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruff==0.0.263
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
toml==0.10.2
tomlkit==0.11.7
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
wrapt==1.15.0
